LPS-151492 Create a Util to export locale constants and use them in all Object Web
