+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/EditObjectValidation.tsx
@@ -17,6 +17,7 @@ import ClayTabs from '@clayui/tabs';
 import {fetch} from 'frontend-js-web';
 import React, {useState} from 'react';
 
+import {availableLocales, defaultLocale} from '../utils/locale';
 import {BasicInfo, Conditions} from './DataValidation/ObjectValidationTabs';
 import {useObjectValidationForm} from './ObjectValidationFormBase';
 import SidePanelContent from './SidePanelContent';
@@ -31,23 +32,6 @@ const TABS = [
 		label: Liferay.Language.get('conditions'),
 	},
 ];
-
-const defaultLanguageId = Liferay.ThemeDisplay.getDefaultLanguageId() as Locale;
-const defaultSymbol = defaultLanguageId.replace('_', '-').toLocaleLowerCase();
-const locales: {label: string; symbol: string}[] = [];
-const languageLabels: string[] = [];
-const languages = Liferay.Language.available as LocalizedValue<string>;
-
-Object.entries(languages).forEach(([languageId, label]) => {
-	locales.push({
-		label: languageId,
-		symbol: languageId.replace('_', '-').toLocaleLowerCase(),
-	});
-
-	languageLabels.push(label);
-});
-
-const defaultLocale = locales.find(({symbol}) => symbol === defaultSymbol);
 
 function closeSidePanel() {
 	const parentWindow = Liferay.Util.getOpener();
@@ -124,7 +108,7 @@ export default function EditObjectValidation({
 									disabled={readOnly}
 									errors={errors}
 									handleChange={handleChange}
-									locales={locales}
+									locales={availableLocales}
 									setValues={setValues}
 									values={values}
 								/>
