+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/Layout/LayoutScreen/ModalAddObjectLayoutField.tsx
@@ -19,14 +19,13 @@ import classNames from 'classnames';
 import React, {useContext, useMemo, useState} from 'react';
 
 import useForm from '../../../hooks/useForm';
+import {defaultLanguageId} from '../../../utils/locale';
 import AutoComplete from '../../Form/AutoComplete';
 import LayoutContext, {TYPES} from '../context';
 import {TObjectField} from '../types';
 import RequiredLabel from './RequiredLabel';
 
 const objectFieldSizes = [1, 2, 3];
-
-const defaultLanguageId = Liferay.ThemeDisplay.getDefaultLanguageId();
 
 type TInitialValues = {
 	objectFieldId: number;
