+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/Layout/LayoutScreen/ObjectLayoutField.tsx
@@ -14,13 +14,12 @@
 
 import React, {useContext} from 'react';
 
+import {defaultLanguageId} from '../../../utils/locale';
 import Panel from '../../Panel/Panel';
 import LayoutContext, {TYPES} from '../context';
 import {TObjectField} from '../types';
 import DropdownWithDeleteButton from './DropdownWithDeleteButton';
 import RequiredLabel from './RequiredLabel';
-
-const defaultLanguageId = Liferay.ThemeDisplay.getDefaultLanguageId();
 
 interface IObjectLayoutFieldProps extends React.HTMLAttributes<HTMLElement> {
 	boxIndex: number;
