+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/Layout/LayoutScreen/ObjectLayoutRelationship.tsx
@@ -15,11 +15,10 @@
 import ClayLabel from '@clayui/label';
 import React, {useContext} from 'react';
 
+import {defaultLanguageId} from '../../../utils/locale';
 import Panel from '../../Panel/Panel';
 import LayoutContext from '../context';
 import {TObjectRelationship} from '../types';
-
-const defaultLanguageId = Liferay.ThemeDisplay.getDefaultLanguageId();
 
 interface IObjectLayoutRelationshipProps
 	extends React.HTMLAttributes<HTMLElement> {
