+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/Layout/LayoutScreen/ObjectLayoutTabs.tsx
@@ -18,14 +18,13 @@ import ClayLabel from '@clayui/label';
 import {useModal} from '@clayui/modal';
 import React, {useContext, useState} from 'react';
 
+import {defaultLanguageId} from '../../../utils/locale';
 import Panel from '../../Panel/Panel';
 import LayoutContext, {TYPES} from '../context';
 import DropdownWithDeleteButton from './DropdownWithDeleteButton';
 import ModalAddObjectLayoutBox from './ModalAddObjectLayoutBox';
 import ObjectLayoutBox from './ObjectLayoutBox';
 import ObjectLayoutRelationship from './ObjectLayoutRelationship';
-
-const defaultLanguageId = Liferay.ThemeDisplay.getDefaultLanguageId();
 
 const ObjectLayoutTabs: React.FC<React.HTMLAttributes<HTMLElement>> = () => {
 	const [{isViewOnly, objectLayout}, dispatch] = useContext(LayoutContext);
