+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/ModalAddListTypeEntry.tsx
@@ -20,24 +20,14 @@ import {fetch} from 'frontend-js-web';
 import React, {useEffect, useState} from 'react';
 
 import useForm from '../hooks/useForm';
-import {normalizeLanguageId, toCamelCase} from '../utils/string';
+import {
+	availableLocales,
+	defaultLanguageId,
+	defaultLocale,
+} from '../utils/locale';
+import {toCamelCase} from '../utils/string';
 import Input from './Form/Input';
 import InputLocalized from './Form/InputLocalized/InputLocalized';
-
-const defaultLanguageId: string = normalizeLanguageId(
-	Liferay.ThemeDisplay.getDefaultLanguageId()
-);
-
-const availableLocales: TLocale[] = Object.keys(Liferay.Language.available).map(
-	(language) => {
-		const formattedLocales = language.replace('_', '-');
-
-		return {
-			label: language,
-			symbol: formattedLocales.toLowerCase(),
-		};
-	}
-);
 
 const ModalAddListTypeEntry: React.FC<IProps> = ({
 	apiURL,
@@ -46,7 +36,10 @@ const ModalAddListTypeEntry: React.FC<IProps> = ({
 }) => {
 	const [error, setError] = useState<string>('');
 	const [selectedLocale, setSelectedLocale] = useState<TLocale>(
-		availableLocales[0]
+		defaultLocale as {
+			label: string;
+			symbol: string;
+		}
 	);
 	const initialValues: TInitialValues = {
 		key: undefined,
