+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/ModalAddObjectDefinition.tsx
@@ -21,6 +21,7 @@ import React, {useEffect, useState} from 'react';
 
 import useForm from '../hooks/useForm';
 import {ERRORS} from '../utils/errors';
+import {defaultLanguageId} from '../utils/locale';
 import {
 	firstLetterUppercase,
 	removeAllSpecialCharacters,
@@ -53,8 +54,6 @@ const ModalAddObjectDefinition: React.FC<IProps> = ({
 		pluralLabel: '',
 	};
 	const [error, setError] = useState<string>('');
-
-	const defaultLanguageId = Liferay.ThemeDisplay.getDefaultLanguageId();
 
 	const onSubmit = async ({label, name, pluralLabel}: TInitialValues) => {
 		const response = await fetch(apiURL, {
