+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/ModalBasicWithFieldName.tsx
@@ -20,6 +20,7 @@ import {fetch} from 'frontend-js-web';
 import React, {useState} from 'react';
 
 import useForm from '../hooks/useForm';
+import {defaultLanguageId} from '../utils/locale';
 import Input from './Form/Input';
 
 interface IProps extends React.HTMLAttributes<HTMLElement> {
@@ -33,8 +34,6 @@ interface IProps extends React.HTMLAttributes<HTMLElement> {
 type TInitialValues = {
 	name: LocalizedValue<string>;
 };
-
-const defaultLanguageId = Liferay.ThemeDisplay.getDefaultLanguageId() as Locale;
 
 export function ModalBasicWithFieldName({
 	apiURL,
